requirements.txt
Django==5.1.2
django-allauth==0.63.3
django-environ==0.11.2
razorpay==1.4.2
Pillow==11.0.0
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.1
setuptools==75.1.0


